# BackendMarketApp/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Precios y Promociones API
  description: API para gestionar precios base y promociones de productos por tienda (usando IDs numéricos incrementales donde aplica), y consultar precios finales.
  version: 1.2.0
servers:
  - url: http://localhost:4000/api/v1
    description: Servidor de Desarrollo Local
...
paths:
  /tienda/createStore:
      post:
        tags:
          - Tiendas
        summary: Crea una nueva tienda.
        description: Registra una nueva tienda en el sistema. La BD genera un ID numérico único.
        operationId: createStore
        requestBody:
          description: Nombre de la nueva tienda.
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreInput'
              examples:
                valid:
                  summary: Crear Market D
                  value:
                    nombreTienda: "Market D"
                emptyName:
                  summary: Nombre vacío (inválido)
                  value:
                    nombreTienda: "   "
                missingField:
                  summary: Falta el campo nombreTienda
                  value:
                    otroCampo: "valor"
        responses:
          '201':
            description: Tienda creada exitosamente.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Tienda creada exitosamente.
                    tiendaId:
                      type: integer
                      description: El ID numérico autoincremental de la tienda creada.
                      example: 4
                    nombreTienda:
                      type: string
                      example: "Market D"
                example:
                  message: Tienda creada exitosamente.
                  tiendaId: 4
                  nombreTienda: "Market D"
          '400':
            $ref: '#/components/responses/BadRequest'
          '500':
            $ref: '#/components/responses/InternalServerError'
  /productos/createProduct:
    post:
      tags:
        - Productos
      summary: Crea un nuevo producto.
      description: Registra un nuevo producto en el sistema. La BD genera un ID numérico único.
      operationId: createProduct
      requestBody:
        description: Nombre del nuevo producto.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
            examples:
              valid:
                summary: Crear Pan Integral
                value:
                  nombreProducto: "Pan Integral 500g"
              emptyName:
                summary: Nombre vacío (inválido)
                value:
                  nombreProducto: ""
              missingField:
                summary: Falta el campo nombreProducto
                value: {}
      responses:
        '201':
          description: Producto creado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Producto creado exitosamente.
                  productoId:
                    type: integer
                    description: El ID numérico autoincremental del producto creado.
                    example: 102
                  nombreProducto:
                    type: string
                    example: "Pan Integral 500g"
              example:
                message: Producto creado exitosamente.
                productoId: 102
                nombreProducto: "Pan Integral 500g"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /promotions/definitions:
    post:
      tags:
        - Promociones Definiciones
      summary: Crea una nueva definición de promoción.
      description: Define un nuevo tipo de promoción basado en porcentaje de descuento. La BD genera un ID numérico único para esta definición.
      operationId: createPromotionDefinition
      requestBody:
        description: Datos de la nueva definición de promoción (solo el porcentaje).
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromocionDefinicionInput'
            examples:
              valid:
                summary: Definición válida
                value:
                  porcentajeDescuento: 15.50
              invalidPercentage:
                summary: Porcentaje fuera de rango
                value:
                  porcentajeDescuento: -10.00
              missingField:
                summary: Falta el porcentaje
                value: {}
      responses:
        '201':
          description: Definición de promoción creada exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Promotion definition created successfully.
                  promocionId: # El ID numérico de la DEFINICIÓN
                    type: integer
                    description: El ID numérico autoincremental de la definición creada.
                    example: 1
              example:
                message: Promotion definition created successfully.
                promocionId: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /promotions/applications:
    post:
      tags:
        - Promociones Aplicaciones
      summary: Aplica una promoción existente a una combinación tienda/producto/período.
      description: Crea un registro que indica que una promoción definida (identificada por su ID numérico) aplica a un producto (ID numérico) en una tienda (ID numérico) durante un período específico. Permite solapamientos. La BD genera un ID único para esta aplicación.
      operationId: createPromotionApplication
      requestBody:
        description: Datos de la aplicación de promoción.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromocionAplicacionInput' # Schema de entrada de Aplicación
            examples:
              valid:
                summary: Aplicación válida
                value:
                  tiendaId: 2       # ID Numérico
                  productoId: 101    # ID Numérico
                  promocionId: 1
                  tiempoPVigenciaInicio: "2025-08-15T00:00:00Z"
                  tiempoPVigenciaFin: "2025-08-31T23:59:59Z"
              invalidDateRange:
                summary: Fechas inválidas
                value:
                  tiendaId: 2
                  productoId: 101
                  promocionId: 1
                  tiempoPVigenciaInicio: "2025-09-01T00:00:00Z"
                  tiempoPVigenciaFin: "2025-08-31T23:59:59Z"
              missingField:
                summary: Falta el ID de promoción
                value:
                  tiendaId: 2
                  productoId: 101
                  # promocionId: 1 # Campo faltante
                  tiempoPVigenciaInicio: "2025-08-15T00:00:00Z"
                  tiempoPVigenciaFin: "2025-08-31T23:59:59Z"
              nonExistentFK:
                summary: ID de promoción que no existe
                value:
                  tiendaId: 2
                  productoId: 101
                  promocionId: 9999 # ID de promoción inválido
                  tiempoPVigenciaInicio: "2025-08-15T00:00:00Z"
                  tiempoPVigenciaFin: "2025-08-31T23:59:59Z"
      responses:
        '201':
          description: Aplicación de promoción creada exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Promotion application created successfully.
                  aplicacionId: # El ID de la APLICACIÓN
                    type: integer
                    description: El ID numérico autoincremental del registro de aplicación creado.
                    example: 2015
              example:
                message: Promotion application created successfully.
                aplicacionId: 2015
        '400': # Incluye FKs no existentes
           description: Solicitud inválida (campos faltantes, formato, fechas, o FK inexistente para Tienda/Producto/PromociónDefinición).
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/Error'
               examples:
                 missingField:
                   value: { message: "Missing required fields: promocionId, ..." }
                 fkViolation:
                   value: { message: "Invalid reference ID (Promotion definition with ID '9999' does not exist)." }
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    StoreInput:
      type: object
      required:
        - nombreTienda
      properties:
        nombreTienda:
          type: string
          description: Nombre de la tienda. No puede estar vacío.
          example: "Supermercado Sur"
    ProductInput:
      type: object
      required:
        - nombreProducto
      properties:
        nombreProducto:
          type: string
          description: Nombre del producto. No puede estar vacío.
          example: "Yogur Natural 1L"
...
    PromocionDefinicionInput: # Schema para crear una definición
      type: object
      required: [porcentajeDescuento]
      properties:
        porcentajeDescuento:
          type: number
          format: float
          description: El porcentaje de descuento (e.g., 25.5 para 25.5%). Debe estar entre 0 y 100.
          example: 15.50
    PromocionAplicacionInput: # Schema para crear una aplicación
      type: object
      required: [tiendaId, productoId, promocionId, tiempoPVigenciaInicio, tiempoPVigenciaFin]
      properties:
        tiendaId:
          type: integer
          description: ID numérico de la tienda.
          example: 2
        productoId:
          type: integer
          description: ID numérico del producto.
          example: 101
        promocionId:
          type: integer
          description: ID numérico de la Promocion_Definicion a aplicar.
          example: 1
        tiempoPVigenciaInicio:
          type: string
          format: date-time
          description: Fecha y hora de inicio de la aplicación (ISO 8601 UTC).
          example: "2025-08-15T00:00:00Z"
        tiempoPVigenciaFin:
          type: string
          format: date-time
          description: Fecha y hora de fin de la aplicación (ISO 8601 UTC).
          example: "2025-08-31T23:59:59Z"
...
    Error: # Schema genérico de error
      type: object
      properties:
        message:
          type: string
          description: Descripción del error ocurrido.
      required:
        - message
  responses: # Definiciones genéricas de respuestas
...
    BadRequest:
      description: La solicitud es inválida (faltan campos, formato incorrecto, valores inválidos).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Invalid input data: 'valor' must be a positive number."
    InternalServerError:
      description: Error interno del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Internal Server Error"